Add files via upload
Task:

Your task is to count the total number of customers residing in each country. This will be achieved by performing an INNER JOIN between the 'Customer Demographic' and 'Customer Address' tables, followed by grouping the results by country. The COUNT function will then be used to tally the number of customers in each geographical segment..