Add files via upload
Task:

Your task is to display the name of every customer along with their address, if available. This type of join ensures that all customers from the demographic table are included in the output, even if they do not have a corresponding address entry.

Finally, this query should be LIMITED to 10.